Change templates directory configuration in settings.py
Changed the way the templates directory is configured in the settings.py file of the bookStore application. Originally, the 'DIRS' option in the 'TEMPLATES' setting was defined as a list containing the absolute path of the 'templates' directory. This has been updated to use the new pathlib syntax, where 'DIRS' is now set to `BASE_DIR / 'templates'`, which represents the same 'templates' directory but using the pathlib library for a more modern and cleaner approach.