Change DEBUG setting to False in bookStore/settings.py
The diff shows that the value of the DEBUG variable in the bookStore/settings.py file has been changed from True to False. The DEBUG variable is used to control the debug mode of the Django application. When DEBUG is set to True, detailed error information is displayed in the browser, which can be helpful during development. However, it is not recommended to keep DEBUG set to True in a production environment, as it can expose sensitive information.